refactor: extract admin image manager view
